scripts/training/requirements.txt
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
db-dtypes==1.5.0
filelock==3.23.0
fsspec==2026.2.0
google-api-core==2.29.0
google-auth==2.48.0
google-cloud-bigquery==3.40.1
google-cloud-core==2.5.0
google-cloud-storage==3.9.0
google-crc32c==1.8.0
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
grpcio==1.78.0
grpcio-status==1.78.0
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
MarkupSafe==3.0.3
ml_dtypes==0.5.4
mpmath==1.3.0
networkx==3.4.2
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.4.5
nvidia-nvtx-cu12==12.8.90
onnx==1.20.1
onnx-ir==0.1.16
onnxscript==0.6.2
packaging==26.0
pandas==2.3.3
proto-plus==1.27.1
protobuf==6.33.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.15.3
six==1.17.0
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.10.0
triton==3.6.0
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3